feat: added a go version directive parser for gomodules
A go directive indicates that a module was written assuming the semantics of a given version of Go. The go directive was originally intended to support backward incompatible changes to the Go language (see Go 2 transition). The `go.mod` file can decide to include a version or not, and this change will allow plugins or other consumers of this parser to get a version from a `go.mod` file, or `undefined` in case none exists. (In newer versions of the `go mod` CLI, this defaults to 1.16, but this behavior is not reproduced here).

Alternatively you can decide to parse it in strict mode, which will cause an error to be thrown in case the directive is not correctly defined in the `go.mod` file.